Downloads/PRJCT/backend/requirements.txt
Django==3.2.4
djangorestframework==3.12.4
django-cors-headers==3.7.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
channels==3.0.4
daphne==3.0.2
djangorestframework-simplejwt==5.3.0

django-redis==5.4.0
python-dotenv==1.0.0
websockets==11.0.3
django-cors-headers==4.3.1
requests==2.31.0
redis==5.0.1
psutil==5.9.8  # For system monitoring